Refactored Invoice setting component
It’s too big at the moment
